% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvis_get_obj_func_results.R
\name{scvis_get_obj_func_results}
\alias{scvis_get_obj_func_results}
\title{Get a dataframe of objective function results for each iteration of scvis}
\usage{
scvis_get_obj_func_results(obj_file)
}
\arguments{
\item{obj_file}{Path to the objective function results file after running scvis_train or scvis_map}
}
\description{
Get a dataframe of objective function results for each iteration of scvis
}
