% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvis_train.R
\name{scvis_train}
\alias{scvis_train}
\title{Train a probabilistic parametric mapping}
\usage{
scvis_train(sce, output_dir, sce_assay = "logcounts",
  use_reducedDim = FALSE, reducedDim_name = NULL, config_file = NULL,
  data_label_file = NULL, normalize = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{output_dir}{Path to the directory for outputs}

\item{sce_assay}{The assay of \code{sce} used to obtain the expression values for the calculations}

\item{use_reducedDim}{Whether reducedDim should be used instead of an assay}

\item{reducedDim_name}{The name of the reducedDim that should be used when \code{use_reducedDim = TRUE}}

\item{config_file}{Path to the configuration file (default: scvis/inst/python/lib/scvis/config/model_config.yaml)}

\item{data_label_file}{Path to a one column file (with column header) that provides the corresponding cluster information for each data point, just used for coloring scatter plots (optional)}

\item{normalize}{Positive float number for normalization of expression values (default: maximum expression value)}
}
\description{
Train a probabilistic parametric mapping
}
